[#63][be][user]refactor: 래빗 및 팀원 리뷰 반영
