+++ back/src/docs/asciidoc/user.adoc
@@ -264,7 +264,7 @@ include::{snippets}/user-info-update-204/http-request.adoc[]
 ==== 요청 필드
 include::{snippets}/user-info-update-204/request-fields.adoc[]
 
-==== HTTP 응답 (429 TooManyRequests)
+==== HTTP 응답 (204 NoContent)
 include::{snippets}/user-info-update-204/http-response.adoc[]
 
 ---
